Filter weekly phrdw data filtered by disease type, date, age

get_weekly_phrdw_by_type_date_age() filter the weekly phrdw data by a given disease type, date range, and age range, it performs the following steps:

Filters the weekly phrdw data by the given epiweek date range and age range.

Aggregates the confirmed cases count by the epiweek date.

Selects the date & the corresponding disease type column.

Renames the disease type as confirm.

The input dataframe weekly_phrdw_data must be the output of get_weekly_phrdw().

The input disease type must be either 'sars_cov2', 'rsv', 'flu_a', or 'flu_b'.

The input start date must be earlier than the input end date.

The input start age must be earlier than the input end age.

get_weekly_phrdw_by_type_date_age(
  weekly_phrdw_data,
  type,
  start_date,
  end_date,
  start_age = 0,
  end_age = 150
)

Arguments

weekly_phrdw_data

weekly phrdw data from get_weekly_phrdw(phrdw_flu_daily_count)

type

disease type (e.g. 'sars_cov2', 'rsv', 'flu_a', 'flu_b')

start_date

start date string (e.g. '2022-01-01')

end_date

end date string (e.g. '2022-12-31')

start_age

start age integer (default = 0)

end_age

end age integer (default = 150)

Value

A weekly phrdw data filtered by disease type, date, age

See also

Examples

phrdw_data <- data.frame(
 lis_date_collection = as.Date(c('2022-10-02', '2022-10-02',
                                 '2022-10-09', '2022-10-09',
                                 '2022-10-16', '2022-10-16',
                                 '2022-10-23', '2022-10-23',
                                 '2022-10-30', '2022-10-30')),
 result_lab_name = c('BC Centre for Disease Control','BC Centre for Disease Control',
                     'BC Centre for Disease Control','BC Centre for Disease Control',
                     'BC Centre for Disease Control','BC Centre for Disease Control',
                     'BC Centre for Disease Control','BC Centre for Disease Control',
                     'BC Centre for Disease Control','BC Centre for Disease Control'),
 age_years = c(7,8,7,8,7,8,7,8,7,8),
 sars_cov2 = c(15,17,30,28,41,43,71,74,80,88),
 rsv = c(14,16,29,27,40,42,70,73,79,87),
 flu_a = c(13,15,28,26,39,41,69,72,78,86),
 flu_b = c(12,14,27,25,38,40,68,71,77,85))

weekly_phrdw_data <- get_weekly_phrdw(phrdw_data)

get_weekly_phrdw_by_type_date_age(
                             weekly_phrdw_data = weekly_phrdw_data,
                             type='rsv',
                             start_date='2022-10-01',
                             end_date='2022-11-01',
                             start_age=0,
                             end_age=10)
#> # A tibble: 5 × 2
#>   date       confirm
#>   <date>       <dbl>
#> 1 2022-10-02      30
#> 2 2022-10-09      56
#> 3 2022-10-16      82
#> 4 2022-10-23     143
#> 5 2022-10-30     166