Filter the weekly plover data by date & disease type

get_weekly_plover_by_date_type() filter the weekly plover data by a given disease type and the date range, it performs the following steps:

Filters the weekly plover data by the given epiweek date range.

Selects the epiweek date & flu type columns.

Renames the epiweek date as date & the flu type as confirm.

The input dataframe weekly_plover_data must be the output of get_weekly_plover().

The input flu type must be 'flu_a' or 'flu_b'.

The input start date must be earlier than the input end date.

get_weekly_plover_by_date_type(weekly_plover_data, type, start_date, end_date)

Arguments

weekly_plover_data

weekly plover data from get_weekly_plover()

type

disease type string (e.g. 'flu_a', 'flu_b')

start_date

start date string (e.g. '2022-01-01')

end_date

end date string (e.g. '2022-12-31')

Value

filtered weekly_plover_data

See also

Examples

plover_data <- data.frame(
          epiWeek_date = as.Date(c('2022-10-02', '2022-10-09',
                     '2022-10-16', '2022-10-23', '2022-10-30',
                     '2022-11-06', '2022-11-13', '2022-11-20',
                      '2022-11-27', '2022-12-04')),
          epiWeek_year = c(2022,2022,2022,2022,
                      2022,2022,2022,2022,2022,2022),
           Epiweek = c(40,41,42,43,44,45,46,47,48,49),
           flu_a = c(17,19,32,38,43,45,73,88,94,105),
           flu_b = c(24,31,39,45,50,52,68,83,89,97))

weekly_plover_data <- get_weekly_plover(plover_data)

get_weekly_plover_by_date_type(
                             weekly_plover_data,
                             'flu_a',
                             '2022-10-01',
                             '2022-12-05')
#> # A tibble: 10 × 2
#>    date       confirm
#>    <date>       <dbl>
#>  1 2022-10-02      17
#>  2 2022-10-09      19
#>  3 2022-10-16      32
#>  4 2022-10-23      38
#>  5 2022-10-30      43
#>  6 2022-11-06      45
#>  7 2022-11-13      73
#>  8 2022-11-20      88
#>  9 2022-11-27      94
#> 10 2022-12-04     105