Generate a summary of model validation

Outputs a summary indicating the prediction quantile in which the confirmed weekly case from the data falls. The function warns when a confirmed case data-point is not within the predicted 95% confidence interval

# S3 method for forecast_time_period_epiestim
summary(object, pred_horizon_str = NULL, ...)

Arguments

object

An object of class forecast_time_period_epiestim

pred_horizon_str

A string indicating the prediction horizon time period to plot

...

Additional optional parameters to pass to the summary method

Value

A list containing two dataframes: individual_quantiles and quantiles_summary respectively:

date

Daily or weekly date over which samples were aggregated

data_quantile

Prediction quantile in which the confirmed case from data is located

data_quantile

Data quantiles

count

Number of data points in one of the prediction quantiles (Q1, Q2, Q3, Q4)

Examples

summary(weekly_time_period_result, pred_horizon_str = "1 week ahead")
#> Warning: Prediction percentile intervals do not cover some data-points in validation fits. Some forecasts may not be reliable
#> $individual_quantiles
#> # A tibble: 7 × 7
#> # Groups:   weekly_date [7]
#>   weekly_date coverage                   weighted_diff confirm median.prediction
#>   <date>      <chr>                              <dbl>   <dbl>             <dbl>
#> 1 2022-10-16  only 95 percentile interv…           512      32                16
#> 2 2022-10-23  only 95 percentile interv…           392      38                52
#> 3 2022-10-30  50 and 95 percentile inte…            32      43                39
#> 4 2022-11-06  50 and 95 percentile inte…             2      45                44
#> 5 2022-11-13  Outside 95 percentile int…          1922      73                42
#> 6 2022-11-20  only 95 percentile interv…          2592      88               124
#> 7 2022-11-27  50 and 95 percentile inte…             8      94                96
#> # ℹ 2 more variables: `50 percentile interval` <glue>,
#> #   `95 percentile interval` <glue>
#> 
#> $quantile_summary
#> # A tibble: 3 × 2
#>   coverage                       count
#>   <chr>                          <int>
#> 1 50 and 95 percentile interval      3
#> 2 Outside 95 percentile interval     1
#> 3 only 95 percentile interval        3
#> 
#> $time_weighted_mspe
#> [1] 27.92848
#>