fit_epiestim_model - Function to estimate the reproduction number of an epidemic

A wrapper function for estimate_R from the EpiEstim library to estimate the reproduction number of epidemics to support short-term forecasts

Usage

fit_epiestim_model(
  data,
  dt = 7L,
  type = NULL,
  mean_si = NULL,
  std_si = NULL,
  recon_opt = "match",
  method = "parametric_si",
  mean_prior = NULL,
  std_prior = NULL,
  ...
)

Arguments

data

data frame containing two columns: date and confirm (number of cases)

dt

Integer length of temporal aggregations of the incidence data. This should be an integer or vector of integers. The default value is 7 time units (1 week).

type

character Specifies type of epidemic. Must be one of "flu_a", "flu_b", "rsv", "sars_cov2" or "custom"

mean_si

Numeric User specification of mean of parametric serial interval

std_si

Numeric User specification of standard deviation of parametric serial interval

recon_opt

One of "naive" or "match" to pass on to estimate_R (see help page)

method

One of "non_parametric_si", "parametric_si", "uncertain_si", "si_from_data" or "si_from_sample" to pass on to estimate_R (see help page)

mean_prior

Numeric positive number giving the mean of the common prior distribution for all reproduction numbers

std_prior

Numeric positive number giving the standard deviation of the common prior distribution for all reproduction numbers

...

Other optional parameters to pass on to estimate_R (see help page) to control estimation of reproduction number

Value

Object of class estimate_R (see EpiEstim help page)

Details

fit_epiestim_model currently supports the following epidemics: Influenza, RSV and COVID-19. The default serial intervals for the estimation of R were retrieved from Cowling et al., 2011, Vink et al., 2014 and Madewell et al., 2023 for Influenza A, Influenza B, RSV and COVID (BA.5 Omicron variant) respectively

Examples

fit_epiestim_model(data = weekly_transformed_plover_data, type = "flu_a")
#> Estimated R for iteration: 1
#> Reconstructed incidence for iteration: 1
#> Estimated R for iteration: 2
#> Reconstructed incidence for iteration: 2
#> Estimated R for iteration: 3
#> Reconstructed incidence for iteration: 3
#> Estimated R for iteration: 4
#> Reconstructed incidence for iteration: 4
#> Estimated R for iteration: 5
#> Reconstructed incidence for iteration: 5
#> Estimated R for iteration: 6
#> Reconstructed incidence for iteration: 6
#> Estimated R for iteration: 7
#> Reconstructed incidence for iteration: 7
#> Estimated R for iteration: 8
#> Reconstructed incidence for iteration: 8
#> Estimated R for iteration: 9
#> Reconstructed incidence for iteration: 9
#> Estimated R for iteration: 10
#> Reconstructed incidence for iteration: 10
#> R estimation starts on day 8
#> $R
#>    t_start t_end  Mean(R)    Std(R) Quantile.0.025(R) Quantile.0.05(R)
#> 1        8    14 1.052436 0.2353318         0.6428552        0.6974836
#> 2        9    15 1.082016 0.2365343         0.6691455        0.7244806
#> 3       10    16 1.120932 0.2383663         0.7034305        0.7597080
#> 4       11    17 1.159984 0.2388220         0.7400540        0.7970252
#> 5       12    18 1.196558 0.2374838         0.7772092        0.8345008
#> 6       13    19 1.230066 0.2344380         0.8142374        0.8714674
#> 7       14    20 1.260262 0.2298798         0.8506320        0.9074361
#> 8       15    21 1.286967 0.2240322         0.8858887        0.9419305
#> 9       16    22 1.251614 0.2122306         0.8707070        0.9241476
#> 10      17    23 1.204743 0.1999590         0.8451441        0.8957581
#> 11      18    24 1.162452 0.1897399         0.8207099        0.8689290
#> 12      19    25 1.125608 0.1815147         0.7983251        0.8445848
#> 13      20    26 1.092158 0.1748202         0.7767384        0.8213686
#> 14      21    27 1.060236 0.1692899         0.7547268        0.7979698
#> 15      22    28 1.028378 0.1646723         0.7312774        0.7733132
#> 16      23    29 1.026408 0.1632818         0.7316456        0.7733894
#> 17      24    30 1.032629 0.1630817         0.7380406        0.7798023
#> 18      25    31 1.039705 0.1628965         0.7452481        0.7870375
#> 19      26    32 1.045966 0.1624663         0.7520683        0.7938280
#> 20      27    33 1.051421 0.1617994         0.7584979        0.8001719
#> 21      28    34 1.056314 0.1609409         0.7647012        0.8062445
#> 22      29    35 1.060838 0.1599273         0.7708062        0.8121824
#> 23      30    36 1.055069 0.1580328         0.7683209        0.8092637
#> 24      31    37 1.046779 0.1559528         0.7636809        0.8041308
#> 25      32    38 1.038556 0.1540686         0.7587814        0.7987786
#> 26      33    39 1.030851 0.1524399         0.7539626        0.7935635
#> 27      34    40 1.023543 0.1510430         0.7491449        0.7884001
#> 28      35    41 1.016461 0.1498496         0.7442093        0.7831624
#> 29      36    42 1.009473 0.1488387         0.7390612        0.7777504
#> 30      37    43 1.040468 0.1507743         0.7661603        0.8054935
#> 31      38    44 1.083670 0.1533706         0.8041250        0.8443269
#> 32      39    45 1.129143 0.1551446         0.8457294        0.8866329
#> 33      40    46 1.173310 0.1556496         0.8882439        0.9295536
#> 34      41    47 1.214743 0.1548231         0.9303947        0.9717834
#> 35      42    48 1.252533 0.1527290         0.9711989        1.0123401
#> 36      43    49 1.286015 0.1494962         1.0097987        1.0503847
#> 37      44    50 1.254540 0.1419294         0.9918691        1.0305651
#> 38      45    51 1.209642 0.1338209         0.9616483        0.9982581
#> 39      46    52 1.168475 0.1269844         0.9329074        0.9677391
#> 40      47    53 1.132435 0.1214400         0.9069810        0.9403568
#> 41      48    54 1.099702 0.1169021         0.8825654        0.9147347
#> 42      49    55 1.068528 0.1131359         0.8583404        0.8894912
#> 43      50    56 1.037532 0.1099781         0.8332232        0.8634996
#> 44      51    57 1.030600 0.1086438         0.8287085        0.8586410
#> 45      52    58 1.030207 0.1080372         0.8293845        0.8591720
#> 46      53    59 1.030464 0.1075322         0.8305261        0.8601950
#> 47      54    60 1.030091 0.1069939         0.8311026        0.8606424
#> 48      55    61 1.029134 0.1064276         0.8311520        0.8605534
#> 49      56    62 1.027811 0.1058556         0.8308486        0.8601087
#> 50      57    63 1.026286 0.1052948         0.8303270        0.8594474
#>    Quantile.0.25(R) Median(R) Quantile.0.75(R) Quantile.0.95(R)
#> 1         0.8856326  1.034948         1.200198         1.467056
#> 2         0.9145078  1.064829         1.230810         1.498189
#> 3         0.9523065  1.104082         1.271209         1.539647
#> 4         0.9912427  1.143636         1.310921         1.578719
#> 5         1.0289917  1.180884         1.347054         1.612092
#> 6         1.0648922  1.215204         1.379052         1.639367
#> 7         1.0985528  1.246313         1.406777         1.660679
#> 8         1.1296253  1.273991         1.430173         1.676272
#> 9         1.1026932  1.239639         1.387489         1.619934
#> 10        1.0645329  1.193698         1.332921         1.551407
#> 11        1.0294807  1.152145         1.284195         1.491138
#> 12        0.9984513  1.115866         1.242152         1.439866
#> 13        0.9697205  1.082845         1.204450         1.394721
#> 14        0.9416804  1.051239         1.168989         1.353193
#> 15        0.9130461  1.019602         1.134149         1.313383
#> 16        0.9120740  1.017763         1.131323         1.308920
#> 17        0.9184615  1.024057         1.137457         1.314700
#> 18        0.9256964  1.031210         1.144459         1.321353
#> 19        0.9322894  1.037566         1.150491         1.326759
#> 20        0.9382442  1.043133         1.155568         1.330944
#> 21        0.9437731  1.048152         1.159962         1.334229
#> 22        0.9490422  1.052812         1.163888         1.336872
#> 23        0.9446202  1.047189         1.156933         1.327756
#> 24        0.9378019  1.039045         1.147329         1.315814
#> 25        0.9309099  1.030948         1.137913         1.304291
#> 26        0.9243531  1.023347         1.129173         1.293740
#> 27        0.9180270  1.016122         1.120973         1.283998
#> 28        0.9117820  1.009106         1.113126         1.274848
#> 29        0.9055002  1.002168         1.105486         1.266118
#> 30        0.9351980  1.033194         1.137812         1.300256
#> 31        0.9766632  1.076443         1.182803         1.347667
#> 32        1.0209918  1.122045         1.229560         1.395866
#> 33        1.0649161  1.166435         1.274212         1.440522
#> 34        1.1070451  1.208172         1.315281         1.480120
#> 35        1.1464179  1.246331         1.351889         1.513883
#> 36        1.1822756  1.280227         1.383447         1.541392
#> 37        1.1561186  1.249192         1.347133         1.496759
#> 38        1.1168949  1.204711         1.297015         1.437847
#> 39        1.0805048  1.163878         1.251436         1.384892
#> 40        1.0483324  1.128097         1.211810         1.339311
#> 41        1.0187589  1.095562         1.176133         1.298790
#> 42        0.9902007  1.064538         1.142507         1.261178
#> 43        0.9613881  1.033648         1.109443         1.224811
#> 44        0.9553903  1.026785         1.101652         1.215574
#> 45        0.9554263  1.026433         1.100875         1.214117
#> 46        0.9560418  1.026726         1.100813         1.213485
#> 47        0.9560490  1.026389         1.100098         1.212168
#> 48        0.9554919  1.025468         1.098781         1.210223
#> 49        0.9545712  1.024179         1.097093         1.207902
#> 50        0.9534411  1.022688         1.095210         1.205402
#>    Quantile.0.975(R)
#> 1           1.561334
#> 2           1.592494
#> 3           1.634137
#> 4           1.672768
#> 5           1.704937
#> 6           1.730310
#> 7           1.749131
#> 8           1.761758
#> 9           1.700549
#> 10          1.627085
#> 11          1.562747
#> 12          1.508235
#> 13          1.460489
#> 14          1.416854
#> 15          1.375336
#> 16          1.370285
#> 17          1.375917
#> 18          1.382422
#> 19          1.387583
#> 20          1.391429
#> 21          1.394299
#> 22          1.396465
#> 23          1.386584
#> 24          1.373820
#> 25          1.361558
#> 26          1.350375
#> 27          1.340096
#> 28          1.330494
#> 29          1.321390
#> 30          1.356101
#> 31          1.404274
#> 32          1.452882
#> 33          1.497441
#> 34          1.536428
#> 35          1.569107
#> 36          1.595120
#> 37          1.547599
#> 38          1.485655
#> 39          1.430162
#> 40          1.382538
#> 41          1.340360
#> 42          1.301390
#> 43          1.263906
#> 44          1.254170
#> 45          1.252475
#> 46          1.251643
#> 47          1.250115
#> 48          1.247950
#> 49          1.245410
#> 50          1.242694
#> 
#> $method
#> [1] "parametric_si"
#> 
#> $si_distr
#>           t0           t1           t2           t3           t4           t5 
#> 0.000000e+00 1.155723e-01 3.184104e-01 2.491845e-01 1.500367e-01 8.209107e-02 
#>           t6           t7           t8           t9          t10          t11 
#> 4.270048e-02 2.152480e-02 1.062058e-02 5.159486e-03 2.477068e-03 1.178265e-03 
#>          t12          t13          t14          t15          t16          t17 
#> 5.562911e-04 2.610285e-04 1.218524e-04 5.663395e-05 2.622295e-05 1.210213e-05 
#>          t18          t19          t20          t21          t22          t23 
#> 5.569186e-06 2.556324e-06 1.170725e-06 5.350694e-07 2.441006e-07 1.111746e-07 
#>          t24          t25          t26          t27          t28          t29 
#> 5.055743e-08 2.295956e-08 1.041334e-08 4.717463e-09 2.134806e-09 9.651019e-10 
#>          t30          t31          t32          t33          t34          t35 
#> 4.358924e-10 1.967055e-10 8.870387e-11 3.995578e-11 1.799996e-11 8.088985e-12 
#>          t36          t37          t38          t39          t40          t41 
#> 3.641909e-12 1.628386e-12 7.317924e-13 3.257838e-13 1.427969e-13 6.613599e-14 
#>          t42          t43          t44          t45          t46          t47 
#> 2.240430e-14 2.223777e-14 8.548717e-16 0.000000e+00 0.000000e+00 0.000000e+00 
#>          t48          t49          t50          t51          t52          t53 
#> 9.092727e-15 0.000000e+00 1.654232e-14 6.994405e-16 2.331468e-16 2.331468e-16 
#>          t54          t55          t56          t57          t58          t59 
#> 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 
#>          t60          t61          t62          t63 
#> 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 
#> 
#> $SI.Moments
#>   Mean      Std
#> 1  3.1 1.652061
#> 
#> $dates
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> [51] 51 52 53 54 55 56 57 58 59 60 61 62 63
#> 
#> $I
#>  [1]  2.306486  2.346032  2.386255  2.427168  2.468783  2.511111  2.554165
#>  [8]  2.577838  2.622035  2.666991  2.712717  2.759228  2.806536  2.854655
#> [15]  3.503474  3.810479  4.144388  4.507556  4.902548  5.332153  5.799403
#> [22]  5.283105  5.330868  5.379062  5.427692  5.476762  5.526275  5.576236
#> [29]  5.798319  5.909540  6.022895  6.138424  6.256169  6.376173  6.498479
#> [36]  6.370859  6.390000  6.409199  6.428456  6.447770  6.467142  6.486573
#> [43]  7.992300  8.692656  9.454384 10.282861 11.183937 12.163973 13.229889
#> [50] 12.123415 12.269772 12.417896 12.567809 12.719531 12.873085 13.028492
#> [57] 13.108444 13.213733 13.319866 13.426853 13.534698 13.643410 13.752995
#> 
#> $I_local
#>  [1]  0.000000  2.346032  2.386255  2.427168  2.468783  2.511111  2.554165
#>  [8]  2.577838  2.622035  2.666991  2.712717  2.759228  2.806536  2.854655
#> [15]  3.503474  3.810479  4.144388  4.507556  4.902548  5.332153  5.799403
#> [22]  5.283105  5.330868  5.379062  5.427692  5.476762  5.526275  5.576236
#> [29]  5.798319  5.909540  6.022895  6.138424  6.256169  6.376173  6.498479
#> [36]  6.370859  6.390000  6.409199  6.428456  6.447770  6.467142  6.486573
#> [43]  7.992300  8.692656  9.454384 10.282861 11.183937 12.163973 13.229889
#> [50] 12.123415 12.269772 12.417896 12.567809 12.719531 12.873085 13.028492
#> [57] 13.108444 13.213733 13.319866 13.426853 13.534698 13.643410 13.752995
#> 
#> $I_imported
#>  [1] 2.306486 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [17] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [25] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [33] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [41] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [49] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [57] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> 
#> attr(,"class")
#> [1] "estimate_R"