Filter the weekly plover data by date & disease type

get_weekly_plover_by_date_type() filter the weekly plover data by a given disease type and the date range, it performs the following steps:

Filters the weekly plover data by the given epiweek date range.

Selects the epiweek date & flu type columns.

Renames the epiweek date as date & the flu type as confirm.

The input disease type must be one of 'sars_cov2', 'rsv', 'flu_a', 'flu_b'

The input start date must be earlier than the input end date.

The input dataframe plover_data must have the following columns:

epiWeek_date: Epiweek Date String (e.g. '2019-01-01')

epiWeek_year: Epiweek Year Number (e.g. 2019)

Epiweek: Epiweek Number (e.g. 1, 2, ..., 53)

flu_a: Confirmed Cases Count (e.g. 1, 2, ...)

flu_b: Confirmed Cases Count (e.g. 1, 2, ...)

rsv: Confirmed Cases Count (e.g. 1, 2, ...)

corona: Confirmed Cases Count (e.g. 1, 2, ...)

Usage

get_weekly_plover_by_date_type(
  plover_data,
  type,
  start_date = min(plover_data$epiWeek_date),
  end_date = max(plover_data$epiWeek_date)
)

Arguments

plover_data

raw plover data before any transformation

type

disease type (e.g. 'sars_cov2', 'rsv', 'flu_a', 'flu_b')

start_date

start date string (e.g. '2022-01-01')

end_date

end date string (e.g. '2022-12-31')

Value

filtered weekly_plover_data

Examples

get_weekly_plover_by_date_type(
  plover_data,
  "flu_a",
  "2022-10-01",
  "2022-12-05"
)
#> # A tibble: 10 × 2
#>    date       confirm
#>    <date>       <dbl>
#>  1 2022-10-02      17
#>  2 2022-10-09      19
#>  3 2022-10-16      32
#>  4 2022-10-23      38
#>  5 2022-10-30      43
#>  6 2022-11-06      45
#>  7 2022-11-13      73
#>  8 2022-11-20      88
#>  9 2022-11-27      94
#> 10 2022-12-04     105